requirements.txt
openai>=1.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
pyautogui>=0.9.54
PySide6>=6.6.0

# Optional — install only for the model adapters you use:
# pip install anthropic        # for --model claude
# pip install google-generativeai  # for future Google AI direct support
# pip install pyobjc-framework-Cocoa  # for macOS capture hiding
